WORKS. Fix that copy is longer than needs to be

+++ src/event/ngx_event_openssl.c
@@ -4953,7 +4953,7 @@ ngx_ssl_get_rtt(ngx_connection_t *c, ngx_pool_t *pool, ngx_str_t *s)
 
         return NGX_ERROR;
     }
-    memcpy(s->data, tmp_rtt, sizeof(uint64_t));  // copy the data from memory into variable's data
+    memcpy(s->data, tmp_rtt, sizeof(uint32_t));  // copy the data from memory into variable's data
                                                 // might need to sprintf to a u_char
 
     FILE* rttlogfile = fopen("/tmp/nginx_rtt.log", "a");
